moving tests to JEST from MOCHA
